first part included the acceleration term in the state vector.
In this project we have artificially created a trajectory and added noise to it. The main aim of the project was to determine the original trajectory from the noisy one using constant velocity model of Kalman filter.
 Here we have implemented the filter considering two different cases. 
In the first case we have included the acceleration term in the state vector. This showed a satisfactory result as the range and azimuth standard deviations were reduced sufficiently. But in the other hand we had to compromise with the velocity standard deviation which remains unavoidable.
 In the second case we removed the acceleration term from the state vector. This reduced velocity standard deviation sufficiently but gave patchy results for range and azimuth.